requirements.txt
asgiref==3.4.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==40.0.2
Django==3.2.20
django-cors-headers==3.10.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.2.0
gunicorn==21.2.0
drf-yasg==1.21.7
idna==3.4
importlib-metadata==4.8.3
packaging==21.3
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.4.0
pyparsing==3.0.7
pytz==2023.3
requests==2.27.1
sqlparse==0.4.4
typing_extensions==4.1.1
urllib3==1.26.16
whitenoise==5.3.0
zipp==3.6.0
